+++ src/app/components/navbar/navbarController.js
@@ -338,15 +338,6 @@ angular.module('liveopsConfigPanel')
           });
         }
 
-        if (UserPermissions.hasPermissionInList(PermissionGroups.viewRecordings)) {
-          items.push({
-            label: 'Recordings',
-            stateLink: 'content.recordings',
-            id: 'recording-management-link',
-            order: 4
-          });
-        }
-
         if (UserPermissions.hasPermissionInList(PermissionGroups.viewCustomStats)) {
           items.push({
             label: 'Custom Statistics',